requirements.txt
# Trading Bot with Market Sentiment Analysis
# Python 3.11+

# Market Data & Trading APIs
ccxt>=4.0.0
alpaca-trade-api>=3.0.0
yfinance>=0.2.30

# Sentiment Analysis
vaderSentiment>=3.3.2
textblob>=0.17.1
transformers>=4.30.0
torch>=2.0.0

# Data Sources
tweepy>=4.14.0
praw>=7.7.0
newsapi-python>=0.2.7

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# Storage
redis>=4.5.0

# Scheduling
APScheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.0
